Remove JAlbumDao
